make aphrodite plugin works
